Rename tailwind-components dir to components
