+++ packages/js/src/dashboard/components/notice.js
